マップデータ読み込みを独立、makerにelectron-rendererを指定
