+++ packages/app/src/infra/file/common.ts
@@ -1,0 +1,2 @@
+// なにか書く
+export {}
