+++ packages/maker/src/index.tsx
@@ -3,6 +3,8 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { Store, StoreType } from './State';
 
+import "./ipc";
+
 import "semantic-ui-css/semantic.min.css";
 import './index.css';
 import App from './App';
